+++ src/lib/index.ts
@@ -1,5 +1,5 @@
-export { parseFile, parseProject } from "./parser/index.js";
-export { parseJSDoc } from "./parser/jsdoc.js";
+export { parseFile, parseSource, parseProject } from "./parser/index.js";
+export { parseJSDoc, isJSDocComment } from "./parser/jsdoc.js";
 export { calculateCoverage } from "./analyzer/coverage.js";
 export { generateJSON } from "./generator/json.js";
 export { generateMarkdown } from "./generator/markdown.js";